requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
psutil==6.0.0